requirements.txt
tensorflow==2.21.0
numpy>=1.23.0,<3.0
Pillow>=9.5.0